Store Implementation part 1
Implemented the layout in the store, buttons to buy, scroll view for both player and backgrounds, implemented all the purchasable sprites.

imported all the background sprites in the gameplay scene.

Coins prefab and animation inserted in the gameplay scene, can now spawn and be destroyed on hit.
currency storage completed, currency is now being stored with playerprefs, added a coinText in the gameplay, store and main menu scenes to showcase the total amount of coins.