Store Implementation part 2
When pressing the buy button (players) 700 coins get deducted and are shown in all the scenes, player sprites in gameplay changes (together with animation), the buy button gets disabled to show the equip button.
When clicking the equip button the player can choose which sprite to use, the equip button gets disabled until another equip button is pressed (of a different sprite).

The issue we encountered is with the equip buttons since it resets their "SetActive" as original when leaving and coming back to the store (when clicking to equip the magician the button disappear, but upon coming back it reappears), while the buy button do not reappear anymore.