Store Implementation part 3
Fixed the problems with the characters, now they can be selected, equipped and will spawn in the gameplay scene.
Implemeted background assets in the store and gameplay, they can be bought, equipped and will be shown in the gameplay.
Store keeps track of what was bought.
Buy UI disappears when buying and equip button appears.
---
Player can now move with both WASD and arrows. (after feedbacks)
Applied the icon for the build.
Removed the player and interactability of the menu, turned all the assets into buttons. (after feedbacks)
---
Implemented tutorial assets to pop up when opening the gameplay scene, the code was not implemented yet, will be openable from the option pop up.