Various implementations
Changed the store background and fixed the general look of it, added the highscore and coins in game over scene.
Small color change on the option pop up after implementing a new button and volume slider for effects
---
Implemented the tutorial from option pop up, and in gameplay.
Now the player will see the tutorial pop up when starting the gameplay, and tick the box to not show it again if they want, when closing it the gameplay will begin.

Changed the animation exit time for the change of players from shop.
Tried to implement effects for powerup (failed).