Various fixes 2
Coins now give 5 of currency.
Fixed the sorting layer in gameplay now the backgrounds don't cover the objects.
Lowered the prices in the store.
All the obstacles and collectables have soft collision (OnTriggerEnter2D)
On hit with an enemy the player will have the collider enabled but the damage value to 0, so now they can still pick up the collectables and destroy the enemies without taking damage for the duration of the invincibility.
Implemented particles when collecting coins, fuel and power ups.
fixed the boundaries of the screen in gameplay.