Various fixes 3
resized the text in main menu to cover the gap between the buttons.
Added a small description in the tutorial on what the collectibles are.
Removed 3 enemies and replaced them with a different one, changed also the tutorial assets accordingly.
Changed the color of the slider to match the fuel sprites.
Fixed the problem with background going to default on restart (added a check in store).